+++ packages/swapkit/helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swapkit/helpers
 
+## 1.0.0-rc.18
+
+### Patch Changes
+
+- bump for fix on shiftDecimals
+
 ## 1.0.0-rc.17
 
 ### Patch Changes